StoreItem model deleted -> new fields to Item model added. Add item view added to 'stores' namespace
