+++ server/src/catalogue/search_indexes.py
@@ -15,6 +15,10 @@ class ItemIndex(indexes.SearchIndex, indexes.Indexable):
     description = indexes.CharField(model_attr='description')
     average_rate = indexes.DecimalField(model_attr='average_rate')
     comments_total = indexes.IntegerField(model_attr='comments_total')
+    store = indexes.IntegerField(model_attr='store__id')
+    quantity = indexes.IntegerField(model_attr='quantity')
+    running_out_level = indexes.IntegerField(model_attr='running_out_level')
+    running_out = indexes.BooleanField(model_attr='running_out')
 
     def get_model(self):
         return Item
