+++ server/src/project/settings.py
@@ -70,9 +70,9 @@ INSTALLED_APPS = [
     'haystack',
     'elasticsearch',
     # my apps
-    'stores',
     'catalogue',
     'authentication',
+    'stores',
 ]
 
 REST_FRAMEWORK = {
