+++ server/src/stores/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Store, StoreItem
+from .models import Store
 
 
 class StoreAdmin(admin.ModelAdmin):
@@ -11,14 +11,4 @@ class StoreAdmin(admin.ModelAdmin):
     class Meta:
         model = Store
 
-
-class StoreItemAdmin(admin.ModelAdmin):
-    list_display = ('name', 'store', 'price', 'quantity', 'running_out_level', 'running_out',
-                    'image_url', 'category', 'average_rate', 'comments_total')
-    list_display_links = ('name',)
-    list_filter = ('category', 'store')
-    ordering = ('price', )
-    search_fields = ('name', 'description', 'store')
-
 admin.site.register(Store, StoreAdmin)
-admin.site.register(StoreItem, StoreItemAdmin)
