+++ server/src/stores/models.py
@@ -1,10 +1,6 @@
 from __future__ import unicode_literals
-
 from django.db import models
 from django.core.urlresolvers import reverse
-from catalogue.models import Item
-
-# Create your models here.
 
 
 class Store(models.Model):
@@ -23,17 +19,3 @@ class Store(models.Model):
 
     def __unicode__(self):
         return self.title
-
-
-class StoreItem(Item):
-
-    store = models.ForeignKey(Store, on_delete=models.CASCADE)
-    quantity = models.IntegerField(default=0, blank=True)
-    running_out_level = models.IntegerField(default=10, blank=True)
-
-    @property
-    def running_out(self):
-        """
-        Function returns True if quantity of items in stock is less or equal to running out level set
-        """
-        return self.quantity <= self.running_out_level
