+++ server/src/stores/urls.py
@@ -1,8 +1,9 @@
 from django.conf.urls import url
+from catalogue.views import ItemAddView
 from . import views
 
 urlpatterns = [
-    # url(r'^additem', views.ItemAddView.as_view(), name='add_item'),
+    url(r'^add-item', ItemAddView.as_view(), name='add_item'),
     # url(r'^categoryadd/$', views.CategoryAddView.as_view(), name='add_category'),
     # url(r'^categorylist/(?P<pk>\d+)$', views.CategoryListView.as_view(), name='category_list'),
     # url(r'^categorylist/$', views.CategoryListView.as_view(), name='category_list_root'),
