+++ server/src/templates/search/indexes/catalogue/item_text.txt
@@ -1,3 +1,5 @@
 {{ object.price }}
 {{ object.name }}
 {{ object.category }}
+{{ object.store }}
+{{ object.average_rate }}
