--- server/src/templates/search/indexes/stores/storeitem_text.txt
@@ -1,5 +1,0 @@
-{{ object.price }}
-{{ object.name }}
-{{ object.category }}
-{{ object.store }}
-{{ object.average_rate }}
